code refactor + fix home screen event card issue
