+++ NAMI/Events/Calendar/CalendarDayCell.swift
@@ -10,20 +10,20 @@ struct CalendarDayCell: View {
     let events: [Event]
     let isCurrentMonth: Bool
     let isSelected: Bool
-
+    
     private let calendar = Calendar.current
-
+    
     //Maximum no event indicators to show
     private let maxEventIndicators = 3
-
+    
     private var dayNumber: Int {
         calendar.component(.day, from: date)
     }
-
+    
     private var isToday: Bool {
         calendar.isDateInToday(date)
     }
-
+    
     var body: some View {
         VStack(spacing: 4) {
             //Day number
@@ -34,7 +34,7 @@ struct CalendarDayCell: View {
                 .frame(width: 32, height: 32)
                 .background(backgroundShape)
                 .animation(.snappy(duration: 0.1), value: isSelected)
-
+            
             //Event indicator
             if !events.isEmpty {
                 HStack(spacing: 4) {
@@ -43,7 +43,7 @@ struct CalendarDayCell: View {
                             .fill(event.eventCategory.color)
                             .frame(width: 6, height: 6)
                     }
-
+                    
                     if events.count > maxEventIndicators {
                         Text("+\(events.count - maxEventIndicators)")
                             .font(.system(size: 8))
@@ -59,7 +59,7 @@ struct CalendarDayCell: View {
         .frame(height: 45)
         .contentShape(Rectangle())
     }
-
+    
     private var foregroundColor: Color {
         if !isCurrentMonth {
             return .gray.opacity(0.5)
@@ -72,7 +72,7 @@ struct CalendarDayCell: View {
         }
         return .primary
     }
-
+    
     @ViewBuilder
     private var backgroundShape: some View {
         if isSelected {
@@ -85,13 +85,11 @@ struct CalendarDayCell: View {
     }
 }
 
-// Preview
-
 struct CalendarDayCell_Previews: PreviewProvider {
     static var previews: some View {
         let today = Date()
         let calendar = Calendar.current
-
+        
         Group {
             //Today with no events
             CalendarDayCell(
@@ -100,7 +98,7 @@ struct CalendarDayCell_Previews: PreviewProvider {
                 isCurrentMonth: true,
                 isSelected: false
             )
-
+            
             // Selected day with events
             CalendarDayCell(
                 date: calendar.date(byAdding: .day, value: 1, to: today)!,
@@ -108,7 +106,7 @@ struct CalendarDayCell_Previews: PreviewProvider {
                 isCurrentMonth: true,
                 isSelected: true
             )
-
+            
             // Non-current month day
             CalendarDayCell(
                 date: calendar.date(byAdding: .month, value: 1, to: today)!,
@@ -116,7 +114,7 @@ struct CalendarDayCell_Previews: PreviewProvider {
                 isCurrentMonth: false,
                 isSelected: false
             )
-
+            
         }
         .previewLayout(.sizeThatFits)
         .padding()
