+++ NAMI/Events/EventHelpers/EventModel.swift
