+++ NAMI/Events/EventHelpers/EventsManager.swift
@@ -32,25 +32,25 @@ class EventsManager {
             return false
         }
     }
-
+    
     func deleteEventFromDatabase(eventId: String) {
         db.collection("events").document(eventId).delete()
     }
-
+    
     func registerUserForEvent(eventId: String, userId: String) {
         db.collection("events").document(eventId).updateData(["registeredUsersIds": FieldValue.arrayUnion([userId])])
         db.collection("users").document(userId).updateData(["registeredEventsIds": FieldValue.arrayUnion([eventId])])
     }
-
+    
     func cancelRegistrationForEvent(eventId: String, userId: String) {
         db.collection("events").document(eventId).updateData(["registeredUsersIds": FieldValue.arrayRemove([userId])])
         db.collection("users").document(userId).updateData(["registeredEventsIds": FieldValue.arrayRemove([eventId])])
     }
-
+    
     func uploadImageToStorage(image: UIImage?) async -> String {
         guard let imageData = image?.jpegData(compressionQuality: 1) else { return "" }
         let ref = storage.reference().child("eventImages/\(UUID().uuidString).jpg")
-
+        
         do {
             let _ = try await ref.putDataAsync(imageData)
             let imageURL = try await ref.downloadURL()
@@ -61,7 +61,7 @@ class EventsManager {
             return ""
         }
     }
-
+    
     func deleteImageFromStorage(imageURL: String) {
         Task {
             do {
