+++ NAMI/Events/EventHelpers/EventsViewRouter.swift
@@ -9,25 +9,25 @@ import SwiftUI
 
 @Observable
 final class EventsViewRouter {
-
+    
     public enum Destination: Hashable {
         case eventDetailView(event: Event)
         case eventCreationView(event: Event)
         case eventUpdateView(event: Event)
     }
-
+    
     var navPath = NavigationPath()
-
+    
     func navigate(to destination: Destination) {
         navPath.append(destination)
     }
-
+    
     func navigateBack() {
         if navPath.count > 0 {
             navPath.removeLast()
         }
     }
-
+    
     func navigateToRoot() {
         navPath.removeLast(navPath.count)
     }
